Fix starting pose for non-holonomic reversed paths.
This commit fixes a bug where the commanded velocities will try to turn
the robot around by 180 degrees when running a reversed path.

The way reversed path support currently works is that we pretend the
robot is rotated 180 degrees when creating the path, then when following
the path we rotated again by 180, as well as invert the commanded
velocities. However, before this commit, in the path following command,
the current pose of the robot was not flipped, which would force users
to flip their pose whenever running a reversed path.

Simply creating reversed path where the starting/ending rotation is
flipped 180 degrees doesn't work because then robot will turn twice
(once at the beginning and once at the end). Plus it doesn't match the
paths in the pathplanner GUI.

An additional fix is potentially required for fixing replanning.